# docs/swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Groccery API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://groccery-api.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/category/{id}": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Category"],
        "summary": "Get a category with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Category ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Product ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "xAuth": [] }],
        "tags": ["Category"],
        "summary": "Delete a category with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Category ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Category deleted successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },

          "400": {
            "description": "Invalid Category ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [{ "xAuth": [] }],
        "tags": ["Category"],
        "summary": "Update a product",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Category ID",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "default": "Fruits" },
                  "image": {
                    "type": "file",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": { "type": "string", "default": "success" },
                "message": {
                  "type": "string",
                  "default": "Product updated successfully."
                },
                "data": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Product ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "summary": "Get a list of products",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "count": { "type": "number" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "consumes": ["multipart/form-data"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "default": "Red strawberry" },
                  "category": {
                    "type": "string",
                    "default": "63705b8724a507ee30966816"
                  },
                  "brand": { "type": "string", "default": "Red strawberry" },
                  "description": {
                    "type": "string",
                    "default": "Fresh Red strawberry & blue berries"
                  },
                  "weight": { "type": "string", "default": "100 Gm" },
                  "salePrice": { "type": "number", "default": 100 },
                  "retailPrice": { "type": "number", "default": 90 },
                  "rating": { "type": "number", "default": 4.3 },
                  "stock": { "type": "number", "default": 10 },
                  "image": {
                    "type": "file",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": { "type": "string", "default": "success" },
                "message": {
                  "type": "string",
                  "default": "Product added successfully."
                },
                "data": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validaton error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products/hot": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "summary": "Get a list of hot products",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "count": { "type": "number" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "summary": "Get a product with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Product ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Product ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "summary": "Delete a product with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Product ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Product deleted successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "summary": "Update a product",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Product ID",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "default": "Red strawberry" },
                  "category": {
                    "type": "string",
                    "default": "63705b8724a507ee30966816"
                  },
                  "brand": { "type": "string", "default": "Red strawberry" },
                  "description": {
                    "type": "string",
                    "default": "Fresh Red strawberry & blue berries"
                  },
                  "weight": { "type": "string", "default": "100 Gm" },
                  "salePrice": { "type": "number", "default": 100 },
                  "retailPrice": { "type": "number", "default": 90 },
                  "rating": { "type": "number", "default": 4.3 },
                  "stock": { "type": "number", "default": 10 },
                  "image": {
                    "type": "file",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Product updated successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Product ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/products/category/{id}": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Product"],
        "summary": "Get a product list by category",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "category ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "count": { "type": "number" },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Product ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "xAuth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "category": { "type": "string" },
          "brand": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" },
          "weight": { "type": "string" },
          "salePrice": { "type": "number" },
          "retailPrice": { "type": "number" },
          "rating": { "type": "number" },
          "reviews": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" },
          "_id": { "type": "string" }
        }
      },
...
      "Category": {
        "properties": {
          "title": { "type": "string" },
          "image": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" },
          "_id": { "type": "string" }
        }
      },
...
      "Error": {
        "properties": {
          "status": { "type": "string", "default": "error" },
          "message": { "type": "string" }
        }
      }
    }
  }
}